Fix various warnings (#468)
* cpu: Fix warnings

* app: Fix warnings

* modules/SceAppUtil: Fix warnings

* modules/SceCtrl: Fix warnings

* renderer: Fix warnings

* gui: Fix warnings

* io: Fix warnings

* kernel: Fix warnings

* mem: Fix warnings

* modules/SceDisplay: Fix warnings

* modules/SceLibKernel: Fix warnings

* modules/SceProcessmgr: Fix warnings

* modules/sceRtcUser: Fix warnings

* modules/SceSysmodule: Fix warnings

* screen_renderer: Fix warnings

* init: Fix warnings

* config: Fix warnings

* gxp/textures: Fix warnings

* modules/SceLibc: Fix warnings

* net: Fix warnings

* formatting: Run clang-format

+++ src/emulator/gui/src/gui.cpp
@@ -120,7 +120,7 @@ static void init_font(State &gui) {
     ImGuiIO &io = ImGui::GetIO();
     ImFontConfig font_config{};
     gui.monospaced_font = io.Fonts->AddFontDefault();
-    gui.normal_font = io.Fonts->AddFontFromMemoryTTF(gui.font_data.data(), font_file_size, 16, &font_config, io.Fonts->GetGlyphRangesJapanese());
+    gui.normal_font = io.Fonts->AddFontFromMemoryTTF(gui.font_data.data(), static_cast<int>(font_file_size), 16, &font_config, io.Fonts->GetGlyphRangesJapanese());
 }
 
 static void init_background(State &gui, const std::string &image_path) {
