+++ src/emulator/modules/SceLibc/SceLibm.cpp
@@ -75,9 +75,9 @@ EXPORT(int, _FSinh) {
 
 EXPORT(int, _FSinx, float a, unsigned int tag, int c) {
     if (tag == 1)
-        return cosf(a);
+        return static_cast<int>(cosf(a));
     else
-        return sinf(a);
+        return static_cast<int>(sinf(a));
 }
 
 EXPORT(int, _Fpcomp) {
